last changes

+++ admin/includes/view_all_posts.php
@@ -46,11 +46,40 @@ confirmQuery($update_to_delete_status );
 
 break;
 
-  }
 
+case 'clone':
+
+    $query = "SELECT * FROM posts WHERE post_id = '{$postValueId}' " ;
+                                $select_post_query = mysqli_query($connection, $query);
+                                
+                                    while ($row = mysqli_fetch_array($select_post_query)){
+                                    $post_title = $row ['post_title'];
+                                    $post_category_id = $row ['post_category_id'];
+                                    $post_date = $row ['post_date'];
+                                    $post_author = $row ['post_author'];
+                                    $post_status = $row ['post_status'];
+                                    $post_image = $row ['image'];
+                                    $post_tags = $row ['post_tags'];
+                                    $post_content  = $row ['post_content'];
+                                                                    
+                                                            
+                                 }
+
+                                 $query = "INSERT INTO posts (post_category_id, post_title, post_author, post_date, image, post_content, post_tags, post_status)";
+                                 $query .= "VALUES ('{$post_category_id}','{$post_title}','{$post_author}',now(),'{$image}','{$post_content}','{$post_tags}', '{$post_status}')";
+
+                                 $copy_query = mysqli_query($connection, $query);
+
+                                 if(!$copy_query){
+
+                                    die("Failed query!");
+                                 }
+
+                                 break; 
 
 }
 
+}
 }
 
 ?>
@@ -68,6 +97,7 @@ break;
                             <option value="publish">Publish</option>
                             <option value="draft">Draft</option>
                             <option value="delete">Delete</option>
+                            <option value="clone">Clone</option>
 
 
 
@@ -105,6 +135,7 @@ break;
                                     <th>View post</th>
                                     <th>Edit</th>
                                     <th>Delete</th>
+                                    <th>Post views</th>
                                 </tr>
                             </thead>
                       
@@ -113,7 +144,7 @@ break;
                            
                            
                            global $connection;
-                                        $query = 'SELECT * FROM posts ';//you can also limit how many categories you want to display by adding "LIMIT n " 
+                                        $query = "SELECT * FROM posts ORDER BY post_id DESC ";//you can also limit how many categories you want to display by adding "LIMIT n " 
                                 $select_posts = mysqli_query($connection, $query);
                                 
                                     while ($row = mysqli_fetch_assoc($select_posts)){
@@ -126,6 +157,7 @@ break;
                                     $post_tags = $row ['post_tags'];
                                     $post_comment_count = $row ['post_comment_count'];
                                     $post_date = $row ['post_date'];
+                                    $post_views_count = $row ['post_views_count'];
                                    
 
                                      echo "<tr>";
@@ -173,7 +205,9 @@ break;
                                      echo "<td>  $post_date </td>";
                                      echo "<td> <a href='../post.php?p_id={$post_id}'>VIEW POST</a> </td>";
                                      echo "<td> <a href='posts.php?source=edit_post&p_id={$post_id}'>EDIT</a> </td>";
-                                     echo "<td> <a href='posts.php?delete={$post_id}'>DELETE</a> </td>";
+                                     echo "<td> <a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='posts.php?delete={$post_id}'>DELETE</a> </td>";
+                                     
+                                     echo "<td><a href='posts.php?reset={$post_id}'> {$post_views_count} </a> </td>";
                                      echo "</tr>";
                                     }
                            
@@ -197,6 +231,15 @@ break;
                             $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
                             $delete_query = mysqli_query($connection, $query);
                             header("Location: posts.php");//this function refresshes page for you, so after pressing delete you dont have to do that because info just disapears correctly
+
+                        }
+
+
+                            if(isset($_GET['reset'])){
+                            $the_post_id = $_GET['reset'];
+                            $query = "UPDATE posts set post_views_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset']) . "";
+                            $reset_query = mysqli_query($connection, $query);
+                            header("Location: posts.php");
 
                         }
